+++ src/routes/api/payments_payku/+server.ts
@@ -5,18 +5,14 @@ import {
   VITE_SANITY_API_WRITE_TOKEN as tokenWithWriteAccess,
   RESEND_API_KEY,
 } from "$env/static/private";
-import {
-  PUBLIC_TOKEN_PAYKU,
-  PUBLIC_PAYKU_API_URL_SANDBOX,
-  PUBLIC_TOKEN_PAYKU_SANDBOX,
-} from "$env/static/public";
+import { PUBLIC_TOKEN_PAYKU, PUBLIC_PAYKU_API_URL } from "$env/static/public";
 import {
   getSanityServerClient,
   overlayDrafts,
 } from "$lib/config/sanity/client";
 import { Resend } from "resend";
 import groq from "groq";
-import UbicationModalTickets from "$lib/components/events/UbicationModalTickets.svelte";
+import { LocaleConfig } from "$lib/utils";
 
 const projectId = import.meta.env.VITE_SANITY_PROJECT_ID;
 const datasetName = import.meta.env.VITE_SANITY_DATASET;
@@ -40,16 +36,6 @@ export const POST: RequestHandler = async (event) => {
 
   const payment_id = body.payment_key;
 
-  const Query = groq`
-  *[_type == "event" && _id == $merchantOrderId && active == true]{
-    _id,
-    slug,
-    ticket,
-    total_tickets,
-    venue,
-  }
-`;
-
   const prePayment = await client.payment.findUnique({
     where: {
       payment_id_service: payment_id,
@@ -62,12 +48,12 @@ export const POST: RequestHandler = async (event) => {
 
   try {
     const response = await fetch(
-      `${PUBLIC_PAYKU_API_URL_SANDBOX}/transaction/${payment_id}`,
+      `${PUBLIC_PAYKU_API_URL}/transaction/${payment_id}`,
       {
         method: "GET",
         headers: {
           "Content-Type": "application/json",
-          Authorization: `Bearer ${PUBLIC_TOKEN_PAYKU_SANDBOX}`,
+          Authorization: `Bearer ${PUBLIC_TOKEN_PAYKU}`,
         },
       }
     );
@@ -85,9 +71,21 @@ export const POST: RequestHandler = async (event) => {
       },
     });
 
+    const eventId = result?.payment?.additional_parameters.event_id || "";
+    const Query = groq`
+        *[_type == "event" && _id == $merchantOrderId && active == true]{
+          _id,
+          slug,
+          ticket,
+          total_tickets,
+          venue,
+        }
+      `;
     const nextEvent = await getSanityServerClient(false).fetch(Query, {
-      merchantOrderId: result.payment.additional_parameters.event_id,
+      merchantOrderId: eventId,
     });
+
+    console.log(nextEvent, "nextEvent");
 
     let ticket = subtractObjects(
       nextEvent[0].ticket.ubication,
@@ -97,16 +95,28 @@ export const POST: RequestHandler = async (event) => {
     // MUTATION PARA ACTUALIZAR EL STOCK DEL STUDIO
     if (result.status === "success") {
       // Enviamos email de confirmación
-      let email = (async function () {
+      (async function () {
         try {
           const data = await resend.emails.send({
             from: "5 Luchas Clandestino <hola@5luchas.cl>",
             to: paymentWithProduct.customer_email,
             // to: 'alejandro.saez@rendalomaq.com',
             subject: `✅ Tú adhesión para ${paymentWithProduct.product.name} fue existosa!`,
-            html: `Hola ${paymentWithProduct.customer_name}, </br> ¡Tu adhesión fue existosa!</br></br> ${paymentWithProduct.ticketAmount} ${paymentWithProduct.ticketAmount > 1 ? 'entradas' : 'entrada'} '${paymentWithProduct.ticketsType}' para ${paymentWithProduct.product.name}.</br></br>Nos vemos en ${nextEvent[0].venue.venueName} | ${nextEvent[0].venue.venueAdress} </br></br>
-            Tú numero de orden es <strong>${paymentWithProduct.payment_id_service}</strong> (No se lo compartas a nadie, te lo pediremos al ingresar) </br></br> <strong>¡Gracias por tu apoyo!</strong> </br></br> <strong>5 Luchas Clandestino</strong>
-            `,
+            html: `Hola ${
+              paymentWithProduct.customer_name
+            }, </br> ¡Tu adhesión fue existosa!</br></br> ${
+              paymentWithProduct.ticketAmount
+            } ${
+              paymentWithProduct.ticketAmount > 1 ? "entradas" : "entrada"
+            } '${paymentWithProduct.ticketsType}' para ${
+              paymentWithProduct.product.name
+            }.</br></br>Nos vemos en ${nextEvent[0].venue.venueName} | ${
+              nextEvent[0].venue.venueAdress
+            } </br></br>
+              Tú numero de orden es <strong>${
+                paymentWithProduct.payment_id_service
+              }</strong> (No se lo compartas a nadie, te lo pediremos al ingresar) </br></br> <strong>¡Gracias por tu apoyo!</strong> </br></br> <strong>5 Luchas Clandestino</strong>
+              `,
             tags: [
               {
                 name: "category",